fix musicpage and lyricpage (#13)
* can not initialize app with firebase db

* add layout lyric and music page (#10)

* fix build failed dimens.xml

* fix firebase

* get data channel from api, store in firestore and parse to artist fragment

* fix musicpage and lyricpage

---------
