+++ app/src/main/AndroidManifest.xml
@@ -26,11 +26,6 @@
             android:exported="true"
             android:theme="@style/AppTheme.Splash"
             >
-<!--        <activity-->
-<!--            android:name=".ui.test.ChannelActivity"-->
-<!--            android:configChanges="orientation|screenSize"-->
-<!--            android:exported="true"-->
-<!--            >-->
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
