+++ app/src/main/java/com/example/android/adapter/FragmentPageAdapter.kt
@@ -4,10 +4,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.example.android.ui.GetStartedFragment
-import com.example.android.ui.bottom_nav_fragment.ArtistFragment
-import com.example.android.ui.login.LoginFragment
-import com.example.android.ui.login.RegisterFragment
+import com.example.android.ui.fragment.login.GetStartedFragment
+import com.example.android.ui.fragment.bottom_nav.ArtistFragment
+import com.example.android.ui.fragment.login.LoginFragment
+import com.example.android.ui.fragment.login.RegisterFragment
 
 class FragmentPageAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle) :
     FragmentStateAdapter(fragmentManager, lifecycle) {
