+++ app/src/main/java/com/example/android/di/module/AppActivityModule.kt
@@ -2,7 +2,7 @@ package com.example.android.di.module
 
 import com.example.android.di.scope.ActivityScoped
 import com.example.android.ui.activity.MainActivity
-import com.example.android.ui.login.LoginActivity
+import com.example.android.ui.activity.LoginActivity
 import com.example.android.ui.test.ChannelActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
