+++ app/src/main/java/com/example/android/ui/fragment/BaseFragment.kt
@@ -1,4 +1,4 @@
-package com.example.android.ui
+package com.example.android.ui.fragment
 
 import androidx.lifecycle.ViewModelProvider
 import dagger.android.support.DaggerFragment
