+++ app/src/main/java/com/example/android/ui/fragment/login/ContinueFragment.kt
@@ -1,4 +1,4 @@
-package com.example.android.ui.login
+package com.example.android.ui.fragment.login
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,13 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.android.R
 
-class RegisterFragment : Fragment() {
+class ContinueFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_register, container, false)
+        return inflater.inflate(R.layout.fragment_continue, container, false)
     }
+
 }
