+++ app/src/main/java/com/example/android/ui/fragment/playing_music/MusicPageFragment.kt
@@ -1,4 +1,4 @@
-package com.example.android.ui.login
+package com.example.android.ui.fragment.playing_music
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.android.R
 
-class SignUpFragment : Fragment() {
+class MusicPageFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_sign_up, container, false)
+        return inflater.inflate(R.layout.fragment_music_page, container, false)
     }
-
 }
