--- app/src/main/java/com/example/android/ui/login/ContinueFragment.kt
@@ -1,20 +1,0 @@
-package com.example.android.ui.login
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.android.R
-
-class ContinueFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_continue, container, false)
-    }
-
-}
