+++ app/src/main/res/layout/fragment_homepage.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".ui.bottom_nav_fragment.HomepageFragment">
+    tools:context=".ui.fragment.bottom_nav.HomepageFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbarHomepage"
