+++ app/src/main/res/layout/fragment_lyric_page.xml
@@ -1,22 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
-    >
     <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        >
+        android:layout_height="match_parent">
 
         <ImageView
             android:id="@+id/backgroundImageView"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:scaleType="centerCrop"
-            android:src="@drawable/img_singer"
-            />
+            android:src="@drawable/img_singer" />
 
         <ImageView
             android:layout_width="match_parent"
@@ -27,46 +24,50 @@
 
 
     <ImageView
-        android:id="@+id/ic_lyric_page"
+        android:id="@+id/icBack"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/definite_dimen_32"
+        android:layout_marginTop="@dimen/definite_dimen_32"
         android:src="@drawable/back_button"
-        android:layout_marginStart="@dimen/SPACE_33"
-        android:layout_marginTop="@dimen/SPACE_66"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/tv_singer_name"
+        android:id="@+id/tvSingerName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="74dp"
-        android:layout_marginTop="@dimen/SPACE_64"
+        android:layout_marginTop="@dimen/definite_dimen_32"
         android:fontFamily="@font/montserrat"
         android:text="@string/tv_singer_name"
         android:textColor="@color/colorPrimary"
-        android:textSize="@dimen/space_24sp"
-        app:layout_constraintStart_toEndOf="@id/ic_lyric_page"
+        android:textSize="@dimen/text_size_xlarge"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toBottomOf ="@id/tv_singer_name"
-        app:layout_constraintStart_toStartOf="parent"
+        android:layout_marginTop="@dimen/definite_dimen_24"
         app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="@dimen/SPACE_24"
-        >
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tvSingerName">
+
         <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rcvData"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
+            android:layout_height="match_parent" />
     </androidx.core.widget.NestedScrollView>
+
     <include
+        android:id="@+id/item_control_music"
         layout="@layout/item_control_music"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
