+++ app/src/main/res/layout/item_control_music.xml
@@ -1,33 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/bottom_nav_background"
-    android:paddingBottom="20dp"
-    >
+    android:paddingBottom="20dp">
 
     <ImageView
         android:id="@+id/img_circle_singer"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
+        android:layout_marginStart="@dimen/definite_dimen_34"
+        android:layout_marginTop="@dimen/definite_dimen_32"
         android:src="@drawable/img_circle_singer"
         app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginTop="@dimen/definite_dimen_32"
-        android:layout_marginStart="@dimen/SPACE_33"
-        />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/tv_singer_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
+        android:layout_marginStart="@dimen/definite_dimen_20"
         android:layout_marginTop="28dp"
         android:fontFamily="@font/montserrat"
         android:text="@string/tv_singer_name"
         android:textColor="@color/colorPrimary"
-        android:textSize="20sp"
+        android:textSize="@dimen/text_size_large"
         app:layout_constraintStart_toEndOf="@id/img_circle_singer"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -35,100 +33,99 @@
         android:id="@+id/tv_first_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/SPACE_20"
+        android:layout_marginStart="@dimen/definite_dimen_20"
         android:fontFamily="@font/montserrat"
         android:text="@string/tv_first_name"
         android:textColor="@color/colorPrimary"
-        android:textSize="@dimen/space_14sp"
+        android:textSize="@dimen/text_size_small"
         app:layout_constraintStart_toEndOf="@id/img_circle_singer"
         app:layout_constraintTop_toBottomOf="@id/tv_singer_name" />
 
     <ImageView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/definite_dimen_44"
+        android:layout_marginEnd="@dimen/definite_dimen_44"
         android:src="@drawable/ic_heart"
-        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="@dimen/SPACE_44"
-        android:layout_marginEnd="43dp"/>
+        app:layout_constraintTop_toTopOf="parent" />
+
     <androidx.appcompat.widget.AppCompatSeekBar
         android:id="@+id/seekbar"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_first_name"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginStart="@dimen/SPACE_30"
-        android:layout_marginEnd="@dimen/SPACE_33"
+        android:layout_marginStart="@dimen/definite_dimen_30"
         android:layout_marginTop="26dp"
-        />
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="2:07"
-        android:textSize="12sp"
-        android:fontFamily="@font/montserrat"
-        app:layout_constraintTop_toBottomOf="@+id/seekbar"
-        android:textColor="@color/colorPrimary"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginStart="@dimen/SPACE_33"/>
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="2:07"
-        android:textSize="12sp"
-        android:fontFamily="@font/montserrat"
-        app:layout_constraintTop_toBottomOf="@+id/seekbar"
-        android:textColor="@color/colorPrimary"
+        android:layout_marginEnd="@dimen/definite_dimen_30"
         app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginStart="@dimen/SPACE_33"
-        android:layout_marginEnd="@dimen/SPACE_44"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tv_first_name" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/definite_dimen_34"
+        android:fontFamily="@font/montserrat"
+        android:text="2:07"
+        android:textColor="@color/colorPrimary"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/seekbar" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/definite_dimen_34"
+        android:layout_marginEnd="@dimen/definite_dimen_44"
+        android:fontFamily="@font/montserrat"
+        android:text="2:07"
+        android:textColor="@color/colorPrimary"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/seekbar" />
 
     <LinearLayout
         android:id="@+id/control_layout"
-        android:layout_marginTop="36dp"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toTopOf="@+id/seekbar"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:gravity="center"
         android:layout_marginStart="70dp"
+        android:layout_marginTop="36dp"
         android:layout_marginEnd="70dp"
-        >
+        android:gravity="center"
+        android:orientation="horizontal"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/seekbar">
 
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/img_repeat"
             android:layout_weight="1"
-            />
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/previous"
-            android:layout_weight="1"
-            />
+            android:src="@drawable/img_repeat" />
 
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/ic_play_music"
             android:layout_weight="1"
-            />
+            android:src="@drawable/previous" />
+
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/next"
             android:layout_weight="1"
-            />
+            android:src="@drawable/ic_play_music" />
+
         <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/shuffle"
             android:layout_weight="1"
-            />
+            android:src="@drawable/next" />
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:src="@drawable/shuffle" />
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
