+++ app/src/main/res/layout/login_activity.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:focusableInTouchMode="true"
-        tools:context=".ui.login.LoginActivity">
+        tools:context=".ui.activity.LoginActivity">
 
         <FrameLayout
             android:layout_width="wrap_content"
